Use CodegenUtils/MarkerUtils
