+++ iree/compiler/Conversion/LinalgToSPIRV/BUILD
@@ -28,7 +28,6 @@ cc_library(
         "KernelDispatchUtils.cpp",
         "LegalizeNumWorkgroupsFnPass.cpp",
         "LinalgTileAndFusePass.cpp",
-        "MarkerUtils.cpp",
         "MatMulVectorizationTest.cpp",
         "Passes.cpp",
         "SplitDispatchFunctionPass.cpp",
@@ -40,7 +39,6 @@ cc_library(
         "Attributes.h",
         "CooperativeMatrixAnalysis.h",
         "KernelDispatchUtils.h",
-        "MarkerUtils.h",
         "MemorySpace.h",
         "Passes.h",
         "Utils.h",
