+++ iree/compiler/Conversion/LinalgToSPIRV/CMakeLists.txt
@@ -21,7 +21,6 @@ iree_cc_library(
     "Attributes.h"
     "CooperativeMatrixAnalysis.h"
     "KernelDispatchUtils.h"
-    "MarkerUtils.h"
     "MemorySpace.h"
     "Passes.h"
     "Utils.h"
@@ -33,7 +32,6 @@ iree_cc_library(
     "KernelDispatchUtils.cpp"
     "LegalizeNumWorkgroupsFnPass.cpp"
     "LinalgTileAndFusePass.cpp"
-    "MarkerUtils.cpp"
     "MatMulVectorizationTest.cpp"
     "Passes.cpp"
     "SplitDispatchFunctionPass.cpp"
