+++ iree/compiler/Conversion/LinalgToSPIRV/Utils.cpp
@@ -21,7 +21,6 @@
 #include "iree/compiler/Conversion/LinalgToSPIRV/Utils.h"
 
 #include "iree/compiler/Conversion/CodegenUtils/MarkerUtils.h"
-#include "iree/compiler/Conversion/LinalgToSPIRV/MarkerUtils.h"
 #include "iree/compiler/Conversion/LinalgToSPIRV/MemorySpace.h"
 #include "mlir/Dialect/GPU/GPUDialect.h"
 #include "mlir/Dialect/Linalg/IR/LinalgOps.h"
